.vscode.dev/pyproject.toml
[project]
name = "bk-monitor-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apigw-manager==3.1.2",
    "arrow==0.6.0",
    "attrs==24.2.0",
    "babel==2.14.0",
    "betterproto==2.0.0b7",
    "bk-audit==1.0.7",
    "bk-crypto-python-sdk==2.0.1",
    "bk-iam==1.3.6",
    "bk-notice-sdk==1.3.2",
    "bkstorages==1.1.1",
    "blueapps==4.15.4",
    "celery==5.4.0",
    "celery-redbeat==2.2.0",
    "confluent-kafka==2.4.0",
    "coverage==7.6.12",
    "dill==0.3.7",
    "django==3.2.25",
    "django-celery-beat==2.7.0",
    "django-celery-results==2.5.1",
    "django-cors-headers==4.1.0",
    "django-db-connection-pool==1.2.5",
    "django-elasticsearch-dsl==7.4",
    "django-filter==23.5",
    "django-jinja==2.10.2",
    "django-mysql==4.10.0",
    "django-prometheus==2.3.1",
    "django-redis==5.4.0",
    "django-versionlog==1.8.0",
    "djangorestframework==3.15.1",
    "dnspython==2.3.0",
    "drf-yasg==1.21.5",
    "elasticmock==1.8.1",
    "elasticsearch==7.17.12",
    "elasticsearch-curator==7.0.1",
    "elasticsearch-dsl==7.4.1",
    "elasticsearch5==5.5.6",
    "elasticsearch6==6.8.2",
    "fakeredis==2.27.0",
    "flake8==5.0.4",
    "furl==2.1.3",
    "gevent==22.10.2",
    "gitpython==3.1.2",
    "google-auth==2.37.0",
    "googletrans==4.0.0rc1",
    "graphviz==0.19.2",
    "gunicorn==23.0.0",
    "humanize==4.6.0",
    "influxdb==5.3.2",
    "ipython==7.34.0",
    "isort==5.11.5",
    "jinja2==3.1.5",
    "jmespath==1.0.1",
    "kafka-python==1.4.6",
    "kubernetes==18.20.0",
    "luqum==0.13.0",
    "mock==5.1.0",
    "netifaces==0.11.0",
    "networkx==2.6.3",
    "opentelemetry-api==1.11.1",
    "opentelemetry-exporter-otlp==1.11.1",
    "opentelemetry-instrumentation-celery==0.30b1",
    "opentelemetry-instrumentation-dbapi==0.30b1",
    "opentelemetry-instrumentation-django==0.30b1",
    "opentelemetry-instrumentation-elasticsearch==0.30b1",
    "opentelemetry-instrumentation-kafka-python==0.30b1",
    "opentelemetry-instrumentation-logging==0.30b1",
    "opentelemetry-instrumentation-redis==0.30b1",
    "opentelemetry-instrumentation-requests==0.30b1",
    "opentelemetry-sdk==1.11.1",
    "pika==1.3.2",
    "pillow==9.5.0",
    "ply==3.11",
    "prometheus-client==0.17.1",
    "protobuf==3.19.5",
    "psutil==6.1.1",
    "pycryptodome==3.18",
    "pyinstrument==4.6.2",
    "pymysql==1.1.1",
    "pypinyin==0.53.0",
    "pyppeteer==1.0.2",
    "pyproject-flake8==5.0.4.post1",
    "pyroscope-io==0.8.8",
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "pytest-django==4.8.0",
    "pytest-mock==3.14.0",
    "python-consul==1.1.0",
    "python-dotenv==0.21.1",
    "python-magic==0.4.27",
    "pytz==2024.2",
    "pyyaml==6.0.1",
    "redis==4.6.0",
    "requests==2.31.0",
    "requests-toolbelt==1.0.0",
    "schema==0.7.7",
    "selenium==3.141.0",
    "six==1.17.0",
    "sqlparse==0.4.4",
    "supervisor==4.2.5",
    "typing-extensions==4.7.1",
    "ujson==5.7.0",
    "whitenoise==6.5.0",
    "xxhash==3.5.0",
]

[tool.black]
line-length = 120
include = '\.pyi?$'
skip-string-normalization = 'true'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | iam
)/
'''

[tool.isort]
profile = "black"
src_paths = ["bkmonitor", "bklog", "bkmonitor/packages"]
known_third_party = ['Crypto', 'PIL', 'apscheduler', 'arrow', 'babel', 'beanstalkc', 'celery',
    'chain_relative_strategy', 'consul', 'coreapi', 'coreschema', 'croniter', 'django',
    'django_apscheduler', 'django_filters', 'dns', 'elasticsearch5', 'fakeredis', 'furl',
    'gitlab', 'graph_exporter', 'humanize', 'influxdb', 'jinja2', 'kafka', 'kazoo', 'kombu',
    'mako', 'mistune', 'mock', 'mockredis', 'netifaces', 'prometheus_client', 'psutil',
    'pygments', 'pymysql', 'pytest', 'pytz', 'redis', 'requests', 'response_tracker',
    'rest_framework', 'rest_framework_bulk', 'rest_framework_nested', 'schematics',
    'six', 'sqlparse', 'supervisor', 'version_log', 'weixin', 'whitenoise', 'yaml', 'blueapps', 'iam']

[tool.flake8]
ignore = "C901,E203,W503,F405,E402"
max-line-length = 120
max-complexity = 25
format = "pylint"
exclude = "*config*,*migrations*,*.pyc,.git,__pycache__.tox,docs,old,build,dist,*.egg-info,.cache,.eggs"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
